pyproject.toml
[tool.poetry]
name = "fms-vpo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
zeep = "^4.2.1"
fastapi = {extras = ["all"], version = "^0.111.0"}
redis = "^5.0.4"
elastic-apm = "^6.22.2"
dynaconf = "^3.2.5"
lxml = "^5.2.2"
pycpfcnpj = "^1.8"
alembic = "^1.13.1"
sqlmodel = "^0.0.18"
psycopg2-binary = "^2.9.9"
retry = "^0.9.2"
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
ruff = "^0.3.6"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.6"
factory-boy = "^3.3.0"
faker = "^24.11.0"

[tool.pytest.ini_options]
addopts = [
    "--cov=./app",
    "--cov-report=html",
    "--cov-fail-under=50",
    "--color=yes",
    "-vv",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
testpaths = ["tests"]

[tool.ruff]
lint.ignore = ["F403"]
line-length = 79
lint.select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"